+++ app/src/main/java/org/brightmindenrichment/street_care/ui/visit/visit_forms/VisitFormFragment1.kt
@@ -2,38 +2,29 @@ package org.brightmindenrichment.street_care.ui.visit.visit_forms
 
 
 import android.app.DatePickerDialog
-import android.app.DatePickerDialog.OnDateSetListener
+import android.app.TimePickerDialog
 import android.os.Bundle
-import android.util.Log
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.findFragment
-import androidx.fragment.app.setFragmentResultListener
 import androidx.navigation.fragment.findNavController
-import com.google.android.gms.common.api.Status
-import com.google.android.libraries.places.api.Places
-import com.google.android.libraries.places.api.model.Place
-import com.google.android.libraries.places.widget.AutocompleteSupportFragment
-import com.google.android.libraries.places.widget.listener.PlaceSelectionListener
-import com.google.android.libraries.places.widget.model.AutocompleteActivityMode
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-import com.google.firebase.ktx.app
-import org.brightmindenrichment.street_care.BuildConfig
 import org.brightmindenrichment.street_care.R
 import org.brightmindenrichment.street_care.databinding.FragmentVisitForm1Binding
-import org.brightmindenrichment.street_care.ui.visit.data.VisitLog
+import org.brightmindenrichment.street_care.databinding.FragmentVisitForm2Binding
 import org.brightmindenrichment.street_care.util.Extensions
+import java.text.SimpleDateFormat
 import java.util.*
 
 class VisitFormFragment1 : Fragment() {
     private var _binding: FragmentVisitForm1Binding? = null
-    private val binding get() = _binding!!
+
+    val binding get() = _binding!!
     private val sharedVisitViewModel: VisitViewModel by activityViewModels()
+    private val myCalendar: Calendar = Calendar.getInstance()
+    private val myCalendar1: Calendar = Calendar.getInstance()
+    private var displayDateFormat: String = "MM/dd/yyyy"
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -44,101 +35,96 @@ class VisitFormFragment1 : Fragment() {
         return binding.root
 
     }
+
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        onViewStateRestored(savedInstanceState)
-        if (Firebase.auth.currentUser == null) {
-            Extensions.showDialog(
-                requireContext(),
-                view.context.getString(R.string.anonymous_user_title),
-                view.context.getString(R.string.anonymous_user_message),
-                view.context.getString(R.string.ok),
-                view.context.getString(R.string.cancel))
+
+        binding.datePickerActions.setOnClickListener {
+            myCalendar.time = populateCalendarToSelectVisitDate()
         }
-            searchLocation()
-            binding.txtGoToPage2.setOnClickListener {
-                //sharedVisitViewModel.saveVisitLog()
-                //  sharedVisitViewModel.visitLog = VisitLog()
+
+
+
+            binding.timePicker.setOnClickListener {
+                val cal = Calendar.getInstance()
+                val timeSetListner = TimePickerDialog.OnTimeSetListener { _, hour, minute ->
+                    myCalendar.set(Calendar.HOUR_OF_DAY, hour)
+                    myCalendar.set(Calendar.MINUTE, minute)
+                    binding.timePicker.text = SimpleDateFormat("HH:mm").format(myCalendar.time)
+
+                }
+                TimePickerDialog(
+                    context,
+                    timeSetListner,
+                    myCalendar.get(Calendar.HOUR_OF_DAY),
+                    myCalendar.get(Calendar.MINUTE),
+                    false
+                ).show()
+            }
+        binding.txtNext2.setOnClickListener {
+
+            //Adding code to fix Date and Time issue in whenVisit and andWhenVisitTime
+            var time = binding.timePicker.text.toString()
+            sharedVisitViewModel.visitLog.whenVisitTime = time
+            var offset = 0
+            if(time.length > 5){
+                val timeFormat = time.substring(5)
+                time = time.substring(0,5)
+                if(timeFormat.contains("pm", false)){
+                    offset = 12
+                }
+            }
+            if(time.contains(":")) {
+                val splitTime = time.split(":")
+                if (splitTime.size > 1) {
+                    myCalendar.set(Calendar.HOUR_OF_DAY, (splitTime[0].toString().toInt() + offset))
+                    myCalendar.set(Calendar.MINUTE, splitTime[1].toString().toInt()) // getting error when tested with single minute time such as 11:08am.
+                    //displayDate(Extensions.dateToString(myCalendar.time, displayDateFormat))
+                    sharedVisitViewModel.visitLog.date = myCalendar.time
+                }
+            }
+            if (findNavController().currentDestination?.id == R.id.visitFormFragment1){
                 findNavController().navigate(R.id.action_visitFormFragment1_to_visitFormFragment2)
-                sharedVisitViewModel.visitLog.locationmap["city"] = binding.edtCity2.text.toString()
-                sharedVisitViewModel.visitLog.locationmap["state"] = binding.edtState3.text.toString()
-                sharedVisitViewModel.visitLog.locationmap["zipcode"] = binding.edtZipcode5.text.toString()
+            }
 
         }
-
-        binding.selectLocationButton.setOnClickListener {
-            // Navigate to the MapSelectorFragment
-            findNavController().navigate(R.id.action_visitFormFragment1_to_mapSelectorFragment)
-        }
-        binding.txtBack.setOnClickListener {
-
+        binding.txtPrevious2.setOnClickListener {
             findNavController().navigate(R.id.action_visitFormFragment1_to_nav_visit)
         }
-        binding.txtSkip.setOnClickListener {
-
-            findNavController().navigate(R.id.action_visitFormFragment1_to_visitFormFragment2)
+        binding.txtSkip2.setOnClickListener {
+            if (findNavController().currentDestination?.id == R.id.visitFormFragment1) {
+                findNavController().navigate(R.id.action_visitFormFragment1_to_visitFormFragment2)
+            }
         }
-        setFragmentResultListener("addressRequestKey") { requestKey, bundle ->
-            val city = bundle.getString("city")
-            val state = bundle.getString("state")
-            val zipCode = bundle.getString("zipCode")
 
-            binding.edtCity2.setText(city)
-            binding.edtState3.setText(state)
-            binding.edtZipcode5.setText(zipCode)
     }
+
+    private fun populateCalendarToSelectVisitDate(): Date {
+        val date = DatePickerDialog.OnDateSetListener { _, year, month, day ->
+            myCalendar.set(Calendar.YEAR, year)
+            myCalendar.set(Calendar.MONTH, month)
+            myCalendar.set(Calendar.DAY_OF_MONTH, day)
+
+            displayDate(Extensions.dateToString(myCalendar.time, displayDateFormat))
+            //setting the user selected date into object
+            sharedVisitViewModel.visitLog.date = myCalendar.time
         }
 
-    // autocomplete places API Using Fragment
-    private fun searchLocation() {
-        // Initialize the SDK
-        if (!Places.isInitialized()) {
-            Places.initialize(requireContext(), BuildConfig.API_KEY_PLACES)
+        context?.let { it1 ->
+            DatePickerDialog(
+                it1,
+                date,
+                myCalendar.get(Calendar.YEAR),
+                myCalendar.get(Calendar.MONTH),
+                myCalendar.get(Calendar.DAY_OF_MONTH)
+            ).show()
         }
-        // Create a new PlacesClient instance
-        val autocompleteFragment: AutocompleteSupportFragment =
-            childFragmentManager.findFragmentById(R.id.autocomplete_fragment) as AutocompleteSupportFragment
-        autocompleteFragment.setActivityMode(AutocompleteActivityMode.OVERLAY)
-        autocompleteFragment.setPlaceFields(listOf(Place.Field.ID, Place.Field.NAME))
-        // Set up a PlaceSelectionListener to handle the response.
-        autocompleteFragment.setOnPlaceSelectedListener(object : PlaceSelectionListener {
-            override fun onPlaceSelected(place: Place) {
-                // setting the place selected by user into our object
-                sharedVisitViewModel.visitLog.location = place.name
-                sharedVisitViewModel.visitLog.locationmap["street"] = sharedVisitViewModel.visitLog.location
-                var city: String? = null
-                var state: String? = null
-                var zipCode: String? = null
+        return myCalendar.time
+    }
 
-                // Extract address components
-                val addressComponents = place.addressComponents
-                if (addressComponents != null) {
-                    for (component in addressComponents.asList()) {
-                        val types = component.types
-                        when {
-                            types.contains("locality") -> {
-                                city = component.name
-                            }
-
-                            types.contains("administrative_area_level_1") -> {
-                                state = component.shortName
-                            }
-
-                            types.contains("postal_code") -> {
-                                zipCode = component.name
-                            }
-                        }
-                    }
-                }
-                binding.edtCity2.setText(city)
-                binding.edtState3.setText(state)
-                binding.edtZipcode5.setText(zipCode)
-                Log.d("BME", getString(R.string.place, place.name, place.id))
-            }
-            override fun onError(status: Status) {
-                Log.w("BME", "An error occurred: $status")
-            }
-        })
+    private fun displayDate(dateString: String) {
+        binding.datePickerActions.setText(dateString)
     }
 
     override fun onDestroy() {
@@ -146,5 +132,6 @@ class VisitFormFragment1 : Fragment() {
         _binding = null
     }
 }
+
