+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -40,9 +40,7 @@
         android:name="org.brightmindenrichment.street_care.ui.visit.visit_forms.VisitFormFragment1"
         android:label="@string/visit_log"
         tools:layout="@layout/fragment_visit_form1" >
-        <action
-            android:id="@+id/action_visitFormFragment1_to_mapSelectorFragment"
-            app:destination="@id/mapSelectorFragment" />
+
         <action
             android:id="@+id/action_visitFormFragment1_to_visitFormFragment2"
             app:destination="@id/visitFormFragment2" />
@@ -58,6 +56,9 @@
         android:name="org.brightmindenrichment.street_care.ui.visit.visit_forms.VisitFormFragment2"
         android:label="@string/log_your_visit"
         tools:layout="@layout/fragment_visit_form2" >
+        <action
+            android:id="@+id/action_visitFormFragment2_to_mapSelectorFragment"
+            app:destination="@id/mapSelectorFragment" />
         <action
             android:id="@+id/action_visitFormFragment2_to_visitFormFragment3"
             app:destination="@id/visitFormFragment3" />
